Moved downgrade_resource from NMOSUtils into IS04Utils
